[FLINK-26546] Extract Observer Interface
